pyproject.toml
[tool.poetry]
name = "vending-machine-app"
version = "0.1.0"
description = ""
authors = ["Rinradee Sae-lao <rinradee.sae@gmail.com>"]
readme = "README.md"
packages = [{include = "vending_machine_app"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.5"
black = "^22.12.0"
isort = "^5.12.0"
django-rest-framework = "^0.1.0"
flake8 = "^6.0.0"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
